Feat: implement 'not found' route
Along with creating an html that is rendered when an unexistent path is requested, I implemented a button that returns the user to the original file (in case they don't click back in the browser). I checked that when the original file is rendered, its script tag is also ran, but no request is made to the server, which was of some concern, as going back to 'home' entails up-to-date info. P.S. Although there are many ways to return redirect the user to other pages, I just used object window's location property.