Feat: work on FE essentials
Started with the mobile-first approach. Created a simple HTML file containing the main structure of the app, as I see it. Linked a script to that file, which executes a successful post request to the server, sending over a newly created task that is generated with a helper function. Did some makeup work with CSS (SCSS) to make app easier to work with.